Fixed particles updating while their life was 0
Added a template AOE ground target ability and a template self cast ability. These have the core functionality required for these types of abilities so to create a similar type of ability the class can be copy-pasted and modified.

Added a self target Channel ability. This is simply intended to be a filler ability to artificially inflate the energy cost of abilities without requiring a lump sum payment.

Moved some boolean ability fields relating to unit targeting into an instance of UnitSearchParams to simplify the logic. In the same vein, added "soft" queries to the UnitSearchParams to indicate parameters where at least one of them needs to match the request.

Added a SpawnSkeleton ability that is comprised of 3 chained Channels followed by the actual ability. Once implemented fully, this will spawn a skeleton friendly to the caster at the selected location.

Added a Revive method to Units