Increased the max sources available to the SoundPlayer from 16 to 32. In my testing, 16 was reasonably possible to fill up and cause a crash via autoclicking menu icons whereas 32 was more resilient.
Added charges, meditation, and focus. Charges limit the use of abilities to the amount of charges available. When the ability is out of charges the player can meditate and use either their focus or health to regain the charge.

Added UI to display focus, meditation, and charges to the GameFooter

Added a field to GameObject to explicitly state whether the object's texture has been loaded yet. This allows any objects that somehow slipped into the renderer without a loaded texture to be skipped silently. These objects being skipped shouldn't be an issue since the texture loading is generally just a case of a single render frame needing to pass to ensure synchronous opengl action (and not seeing something for 1 frame when it is originally loaded is essentially unnoticeable).

Shored up some bugs related to timed hover tooltips sticking around when the parent object is orphaned unexpectedly.

Started adding some common UI elements that aren't quite worth making their own class to UIHelpers to simplify their creation. The current pipeline I'm using is: components with unique usage get their own UIComponent, components that implement several other components to create common functionality get a creator function in UIHelpers, and components that solely relate to gameplay get a definition in GameUIDefinitions.

Added a very janky rendering option to UIObjects; RenderAfterParent. This option allows the UIObject and all of it's children to be skipped in the render process until it's parent is rendered. This gives the object a background to blend with for transparent UIObjects. This is a rare use-case so this is sufficient for the moment but I don't see this one sticking around forever.

Added a couple new FBO's to the renderer for the purpose of creating a multi-stage render pipeline. Currently they do nothing but take up memory but it's better to have that memory allocated now so that I know what I'm working with.

Added a RenderStateManager ContextManager to the RenderingQueue. This will take flags set in the main program and alter the render pipeline to reflect those flags. Currently the only flag is a GaussianBlur flag which renders the scene onto and FBO. This will be expanded in the future to apply a guassian blur using the intermediate FBO's that have been set up.

Added a smooth pan camera feature and added a pan when a controlled unit's turn starts.

Added action points. Action points allow units to make actions while energy allows units to make movements. Action points will be a much more discrete value for abilities (generally 1 to 4) to make balancing easier. Movement energy will remain as a float to allow for slows and movement increasing abilities to maintain their effectiveness.

Added some handling to automatically determine when the player is in combat. Currently this handling is something like "when a player controlled unit is in vision of the enemy, all enemy and allied units in a radius of 20 will be placed in combat". This will need to be improved but it's a good start.

Changed player controlled units to actually use the internal energy tracker. I don't even know why I had it using the energy bar UI component for this for so long. The handling was already there for AI units for pete's sake.

Added a skeleton to the SpawnSkeleton ability that makes use of prefabs and the entity manager.

Fixed a very old spritesheet bug caused by applying the spritesheet width/height aspect ratio to the vertices of the renderable object. This would cause objects in the render with it to be stretched unnaturally if it was the first object in the queue. Very bizarre and hard to track down.

Added a ShouldDecay option to combo abilities. For abilities using the action point economy I feel it would be best if they do not decay but I will test and see what feels best.

Added a DestroyOnUnload option to entities that allows an entity to be removed when it is unloaded. This will most likely be used for summoned units and units created by features.

Added a way to find a prefab by name. These names will all be hardcoded and essentially just cross referenced with the prefab manually.

Vastly improved user interaction with the GameFooter

Added depth testing for UI elements. Each UI element will be assigned a Z position value based on its position in the UI tree.

Added some skeletons to the graveyard. The combat actually slotted in pretty seamlessly which bodes well for future features.

Changed the Q key to center the camera on the scene's current unit.

Improved performance of the GetPathToPoint tile map function. Previously, an O(n) full list search was being used to determine if a tile would have a circular reference which was causing issues on some large tasks. This has now been changed to be an O(1) search on a hashset. (also it now doesn't fill up the application's ram when it finds a circular reference which was a separate but related issue)

Changed how VisionMap calculates temporary vision to not completely recalculate the vision map for each unit. Now it only calculates the vision for each temporary position and the current team vision is used as a supplement to determine what the AI believes will be visible once they move.

The MoveInRangeOfAbility has been changed to benefit from the previous temporary position vision change as well as use limited samples over multiple tasks to more efficiently calculate a valid move. Checking every single move was proving to take way too long in large combats with negligible benefit. With this method, the set of valid tiles are randomized to remove any biases and then they are sampled up to 50 times over 3 tasks. This amount has been sufficient for the AI so far but it can be increased if necessary since there is more than enough performance to spare in this aspect after these improvements.

Changed the AIDispositions to more rigorously check if they are in a valid state to take an action. (Ie they aren't attacking while dead or out of energy in some fringe cases)

Added a diagram depicting and describing a skill tree implementation.