Changed most instances of "public" to "internal" to hopefully protect a tiny bit against people just decompiling the executable.
Added 3D lighting. These use a new Material class which applies a diffuse and specular map as per the basic 3D lighting equations. The diffuse map has replaced the TextureReference in RenderableObjects. Each BaseObject can be made to recieve lighting via an EnableLighting field.

Added a function to rotate cube coordinates.

Added normal coordinates to 3D models (and all models the engine uses in general) so that lighting can be applied correctly.

Added the ability to click the charge indicator to regain ability charges.

Added a combo indicator to ability icons.

Enabled gamma correction for the rendered game area.

Changed a lot of hot path ForEach's to for loops

Added UnitGroups. The UnitGroup allows the player to collect all of their units into one unit for ease of transportation. If the player enters combat while grouped they will be forced to place all of their units (assuming there is room to do so)

Removed the janky old lighting setup.

Cleaned up some issues with the turn display and added the ability to pan to a unit by clicking on their portrait (if they are in vision)

Cleaned up a lot of functionality around starting and stopping combat and updating unit information properly.

Added high frequency tickable objects. These tick at around double the rate of normal tickable objects and can be used sparingly for situations where a smoother animation is required.

Played around with the camera's FOV a bit but ultimately decided to stay to 90 degrees.

Added a cube map and a skybox. The skybox has its own shader and render function.

Added some skeleton themed abilities to start filling up the ability roster.

Added a GenericSelectGround ability that can be used to create a ground target ability for use outside of unit combat. The only usage of it so far is creating the effect to place units when dissolving the unit group.

Cleaned up some ability handling to be a bit more intuitive to use.

Fixed an issue where having an ability selected when pressing the "venture forth" button would essentially soft lock you.

Added some more stringent tests for whether a unit could actually use an ability in the game footer.

Added a side bar. The usage of this UI element will be to hold buttons to access gameplay relevant menus easily. Currently the sidebar contains buttons for accessing a party information screen and a camping screen. In addition, the side bar can be minimized so that it takes up very little space on the screen.

Changed the unit status bars to now only show the health and shield bars when zoomed out far enough. This provides a very pleasant viewing experience from above as opposed to the unnecessarily large name boxes that couldn't even be read from that distance.

Added a bandit camp feature that takes advantage of the new building class. Buildings are structures that take up multiple tiles and as such require a list of tiles called the tile pattern to determine which tiles they will be taking up.
The bandit camp was used to test some interactions with buildings originating on unloaded maps and buildings spilling over into unloaded maps and the results were very nice. Buildings can be loaded from any of the points in their pattern and overhanging models will be gradually faded out by the shader. An additional note on buildings; buildings can be rotated but only in 60 degree increments. This is due to the tile pattern only being able to maintain resolution in 60 degree increments. It may seem limiting but 6 modes of rotation should be enough for the vast majority of buildings. And if 6 rotations isn't enough we can create a second pattern that is pre-rotated 30 degrees with a new tile pattern and gain an additional 6 rotations.

Changed feature equations to be applied to every single map when new maps are loaded as opposed to only applying to new maps. As such the ApplyToTile function now has a freshGeneration flag to help determine whether the action being done is relevant for the first or second pass of the feature equation.

Added a LoadRadius to feature equations that is currently unused but will be used to specify which feature's we are currently checking when loading a new map. Each map can have a unique load radius which will allow for large features such as roads to always be checked while small features like a bandit camp might only have a load radius of a few map lengths.

Finally implemented a real AffectsMap function for feature equations. When a new point is added to a feature equation the corresponding tile map is calculated based off of the width and height we are expecting and thrown into a hashset. Then it's as simple as checking for a hit in the hashset. This change alone brought the load time for 1000 features from 5 seconds to 1 second. And attempting to load an additional 10k features had only 1 total second of loading as well. So it's safe to say that this method will be plenty efficient going forward.

Fixed 3D texture mapping. I have no idea what I was doing before but I had it implemented completely wrong.

Tweaked the day/night cycle to better account for gamma correction and 3D lighting

Finally solved the debacle of the escape menu backdrop modal. Some change or another along the way made it so the transparency just works now as a UI object which is nice.

The E button can now be used to end your turn.

Added a tent building. This can be used as the basic schematic for buildings going forward.

Changed walls to use 3D models instead of planes. Surprisingly there is no noticeable performance drop when doing so which gives me hope for 3D models being easy to plug and play without worrying about performance issues.

Added a manual garbage collection when tile maps are loaded. There is essentially no difference in load times and the memory leak that would appear to happen when moving through many tile maps is no longer appearing.

Made a very simple optimization in the AIActions that saves a ton of performance. Now, instead of checking all potential tiles when determining what to do, all tiles that are farther away than the range of the ability will be thrown out. This means that melee and low/medium range abilities will only have a handful of valid tiles and the optimal tile can usually be selected. (honestly not sure how I didn't see this the first go around)

Added a healer and utility AI disposition. These will prioritize using their heal and utility ability respectively.

Added some information to be returned after Unit.ApplyDamage is called. This can be useful for a multitude of things. If we have an ability or buff that, for example, does an effect every time we deal a certain amount of damage then we can use these return values to track that.

Added a fatigue mechanic to AI units. They will now dislike doing the same action indefinitely.

Added a Textures class adjacent to the Spritesheets class. It's a bit janky honestly but the easiest way to set this up is to define the texture as a 1x1 spritesheet. This way we reap the benefits of all the work we did getting spritesheets working for basically no extra overhead.

Added a boatload of models and textures.