Added a BitOperation static class. Currently it only holds the GetBit method but eventually there might be some more things to add there.
Changed the SetColor function for GameObjects and up to optionally accept an enum value that describes the nature of the color we're supplying. Currently this is only being used to set the DefaultColor when no flag is provided but there is room for expansion there if the need arises.

Heavily cleaned up UIObject coloring. Now the color is evaluated after events that could reasonably change the color. When being evaluated, the color starts as the base color then checks its state for any flags that have an associated color before selecting the most relevant event to apply the color of.

Added a pattern creation tool to simplify building creation.

Added TimedAnimations that function similarly to property animations except they are ticked from outside of the main thread and render loop. These have a guaranteed ~85 ticks per second instead of the variable ticks per second when using the render thread.
This does come with the drawback that any additions or removals of items that need to be rendered will need to be staged to occur before or after a render cycle. The current method I'm using to stage changes is by adding a local function to the scene's PostTickEvent that makes any required changes.

Changed the day/night cycle, the arrow, and the camera pan animations to TimedAnimations.

Added an "out of combat turn" which resolves every ~5 seconds when out of combat. This provides a simple way to tick down buffs when not in combat although the feature will need to be refined.

Added a lock object back to QueuedLists and started using them properly.

Added the ability's CreateIconHoverEffect function into UIHelpers and generalized it a bit.

Moved selection tiles from tile maps to the tile map controller. This fixes a lot of weird interactions and saves a good chunk of memory.

Added a PostTickEvent and an OnRenderEvent that can be hooked into to scenes.

Added a sucker punch ability that can stun the target.

Added a CastingMethod bitmask to abilities. Abilities can have any number of the casting method bits set to determine the weakness of the ability. Currently the casting method is only being used to determine whether a unit's status makes them unable to use that ability, but, in the future, it can be used to assign attribute bonuses based on the casting methods.

Added a StatusCondition field to buffs. StatusConditions currently map 1:1 to CastingMethods as their counter. They can have other uses in the future though as there are ~25 more bits that can be set. These status conditions are enumerated by the unit to determine what the unit's status is.

Cleaned up how buffs are added and removed from units.

Added TileEffects. These are added to a tile and evaluated when their condition is met.

Moved the information bars in the footer to make some space.

Implemented a lock in the TurnDisplay component when setting units.

Moved several unit operations out of the CombatScene and into the OnTurnStart and OnTurnEnd methods.