Switched everything back to using public access modifiers. Turns out internal does basically nothing for security and you might as well just deal with the fact that the game can and will be pirated if it's popular.
Changed text rendering to use signed distance fields and also to have its own stage in the render pipeline. The text actually looks good now when downscaled which is incredibly useful for several features. Instead of applying the scissor method like what was done for generic objects, the scissor data is passed into the shader with the letter and whether the letter should be rendered is handled in the fragment shader. An additional layer of protection also comes from testing the bounds of the text component when it could potentially be outside of the scissor bounds to not send it to the renderer at all. These two techniques are sufficiently performant for the moment.

Changed the ScrollableArea to propagate its scissor data down to all its children. This is necessary to make the new text handling work properly with scissoring.

Fixed some text component anchoring issues.

Added the { and } characters to the character set. Additionally, the character set now contains room for 256 characters as opposed to 100

Added some handling for UIObjects to propagate a Z value down to all its children. This mostly works but is still breaking in some edge cases.

Added an event log to the footer. Currently the extent of the functionality is that an event can be sent to the event log with a message and a severity level and it will be printed into the log and push other events up and out of the log. Eventually this will feature some interactivity for the player.

Added a GetRandom extension function for lists. This simply returns a random element from the list.

Fixed text rendering in a huge way by disabling mipmap generation when using signed distance fields. Text is now crisp and beautiful at all levels of magnification.

Added ability trees. These replace the ability profile system entirely. The way ability trees work is that they are statically defined and can be accessed via a tree identifier and a node id or node name. This then returns what was essentially an ability profile previously. These ability trees have a base ability and then any number of nodes attached to the EntryPoint node. Later these will contain quest ids and feature hashes and things point them to where to unlock the ability.

Supplementing ability trees, ability loadouts have been added. This is a serializable class that allows the data required to access any ability on an ability tree to be saved into the save state data. Thus all saved unit's abilities can be saved and reloaded at will.
Additionally, loadouts allow us to create preset, thematic ability combos that can be used for wide ranges of scenarios. For example, a "skeleton sorceror" or "skeleton warrior" preset could be made with different abilities easily mapped to both.

Rearranged the footer a fair bit. Now all unit information is on the right side of the footer and the event log is on the left. In my opinion this makes everything easier to parse and provides ample space for the event log. All is subject to change though.

Added right click movement. Right clicking on a tile iniates an unlimited range move command on it and right clicking again cancels the movement. This issues an unlimited range move command in combat as well but the movement will be ended if the unit does not have sufficient energy to continue.

Added a FeatureID and FeatureHash field to units. This allows all units spawned by a feature equation to be mapped exactly pre and post save state. If a unit from a feature was spawned by a save state it will not be spawned from the feature. This will eventually be expanded further to save which units have been killed in features and not allow them to spawn until a certain amount of time in game has passed.

Fixed the assumption I had that hashes were consistent between runs of the application and changed all cases where I was using hashes as unique identifiers to other encoding methods.

Added save states. A save state is a serializable class that takes all the information necessary to recreate the current game state and stores it in an XML file. Then when loaded the XML file is deserialized and the information is applied to the current scene as necessary.

MASSIVELY improved tilemap load time by reusing old tilemap textures. This optimization was so simple that I'm genuinely surprised I missed it earlier. Now loading a new set of tile maps takes less than a second and is hardly noticeable compared to the multi-second load times of the past.

Changed all hardcoded ability declarations for units to ability loadouts. When loading the entity the current ability loadout will be applied to the unit.

Expanded UnitProfiles to have a UnitProfileType to simplify serialization.

Added a placeholder file in the Save folder just to ensure the folder gets copied into the build folder.