Added a ConsistentRandom class to ensure that any randomly generated numbers will always be consistent. This class is just the current .Net random implementation
Fixed the hyper annoying hovering tooltip bug. Now when the underlying object is cleaned up the tooltip is destroyed.

Feebly tried to fix the UI z position issue. I am kicking the can on that one to a smarter version of myself.

Added an ObjectsDrawn property to the renderer. This allows us to potentially spot any memory leaks and things not getting derefernced when they should be.

Added dialogue. It is loaded from an XML file and is displayed in a west of loathing style UI setup. It is quite cool if I do say so myself.

Made it so that a combat encounter started on the edge of the tile cluster will load in the surrounding maps. Additionally, any units in the maps that are being unloaded will be teleported to the party. (I might eventually extend this to them being added to the unit group instead of being unloaded if they are player units)

Added a lazy try catch to just throw out an error in FillInTeamFog. This function is called so often that losing a couple runs to a weird edge case basically means nothing.

Added ExclusiveUIObjects. These take all UI interaction for themselves when present. Currently this is only being used for the dialogue window.

Added an OnRenderEndEvent to the basic scene so that operations can be queued up for the end of a render cycle.

Added a lock object to the ContextManager class

Added a CreateRandom function that serves the purpose of randomizing the file names of the XML files. This makes it slightly harder for people to just peak into things and change it willy nilly.

Added a dev ability to just kill an enemy instantly.

Made it so that ability trees could contain multiple basic abilities.

Added some more nuance to the process of applying an ability loadout to a unit. The loadout now stores the charge information and that gets applied to the unit when loaded. This can be extended to other properties eventually.

Added a parameter to the entity loading process to signal whether the unit should be placed on the tile map or not. This allows us to save and load units into and from a unit group.

Added the ledger system. Certain player interactions will send an appropriate ledger update. This ledger can then be polled for any state information regarding the player interactions that have been made. Ledgers currently exist for dialogue outcomes, feature interactions (such as whether its been cleared, whether any given enemy has been killed, how many enemies must be killed to consider the zone cleared, etc), and quests (these reference data from all the ledgers to determine if a quest objective has been completed or not).
Ledgers get saved into the player save data.

Added examples of the feature ledger state values to the bandit camp feature.

Changed the feature coordinate hashing function to simply return a long value containing the 2 ints

Added quests. A single quests consists of 1 or more quest states that contains 1 or more objectives. Quest objectives work by holding expected return values for ledger state values. If the recieved return value is equal to the objective's expected value then the objective is considered complete.

Added all of the ledger save handling into the SaveState class.

Added a species property to units

Changed the vision map to check a larger area for unit vision instead of calculating all of the vision from a single point. This results in a more uniform looking vision area.

Changed the shader fade out area to fade out the edges of the tilemaps slightly. This makes the world feel a bit less cramped with the map just dropping off into the void at the edges.