Added a UIManager object that handles Z positioning of UI and keeping track of valid objects for event checking. This allows us to avoid scenarios where we're checking 500 objects just to see if one of them has a hover or click effect. Due to depth testing precision there is currently a de-facto limit of 999 top level UI elements and 9999 child elements per top level UI element. More top level UI elements and child elements can be created than this but the depth testing will presumably stop working (not going to test this because it's an edge case I doubt I'll ever reach).
Removed the ActionQueue class. Pretty sure I didn't know what a lock was when I made that.

Significantly lowered the far depth of the camera. 1000 was there to make the pseudo-skybox work and now that there is a legitimate skybox it is not necessary.

Changed the way unit AI works. Instead of dispositions for overarching categories of units each ability will now have a GetAction method that calculates the weight of that ability and finds a suitable target. This will still support things like bloodthirsty and virtuous, it will just have to be manually accounted for on each ability (it won't be too bad since most ability archetypes will be fairly copy-paste)

Changed abilities so that they can only be used once per turn. There might be abilities that go against this rule but in general I think it's a more interesting system than allowing spamming.

Removed the old AbilityProfile class. Any prior json handling will be replaced with serializable xml classes.

Stopped mixing textures on the health and focus bars. This was what was causing the weird bug where they'd phase out of existence randomly (and it doesn't even make them look better at all).

Changed all ledgers to take homogenized state values instead of random bits and pieces of state values.

Added state subscribers. These can be added as a state value and will be checked any time the state is changed. If the value of the changed state is found to match a state subscriber then the state values held by the subscriber will be applied to the state. This is essentially equivalent to how the quest objectives are checked but these subscribers can exist without needing to be tied to a quest.

Started toying with the concept of state instructions. These are magic state values for each respective ledger that do arbitrary actions based on the passed data. With enough of these instructions created the system for internal interactions within the game should be extremely robust and require very minimal hardcoding beyond the initial creation of instructions.

Copied all of the serializable class definitions and their serializers from the engine tools GUI application. Parity will need to be manually maintained between these any time a change is made.

Started the process of changing features to use serialized data. This is quite a large undertaking so there is still a lot of work to be done but the progress is good so far.

Added quests to the save state.

Removed all hardcoded objects from the MenuScene. From this point onwards everything should be able to added via features (except when testing of course).

Created a system for buildings that is similar to the previous UnitProfile system. The UnitProfile system was mostly abandoned since it was found to be very difficult to modify but that doesn't seem like it will be a problem for buildings as they will be mostly static.

Added a lot of finagling to the LoadSurroundingTileMaps method to get unit vision to load properly. For the moment it seems to work and not be a pile of race conditions like it used to be.

Right now units are very limited on what they can be (ie they can be a rotated quad of a predetermined size at a chosen height over a tile). This is very low priority at the moment but unit definitions will need to be expanded eventually to allow: multi tile units, segmented units, maybe some other stuff.

Changed UnitAI to use the new action system for abilities.

Added an automatic feature manager that handles loading and unloading of features based on distance from the currently loaded tilemaps. This system is already pretty robust and most likely won't need to be changed except to better obfuscate feature names later in the development process (currently information that is being acted on for the feature is stored in the filename which is certainly not ideal for a user facing product)