Changed how UIObject events are handled to more closely match WPF event handling since I've been getting used to how that works.
Added a scroll event that can be captured by a UI object if the mouse is within the UIObject's bounds when the scroll occurs. Only one scroll event can be fired per update.

Fixed an issue where the ScrollableArea scrollbar wouldn't initialize properly by reworking things to not recreate the scrollbar unnecessarily.

Changed UIBlocks to not use the MultiTexture handling. That handling might be on the way out at some point.

Changed tooltips to not be created so aggressively close to the user's mouse cursor.

Changed how camera zooming was being handled from being hardcoded into the scene's update function to now being the scroll event of a screen sized transparent ui object. This comes with the useful property of not having to finagle to resolve multiple scroll events firing over different arbitrary object types (ie UIObjects, GameObjects, random hardcoded handling in the update functions, etc).

Halved the size of the render buffer to only accept 7500 objects per render pass. This probably doesn't matter too much but moving less data around is always better and if this ends up being a problem the application will let me know very quickly.

Added Layer, LoadPriority, and DescriptiveName fields to FeatureEquations. The layer determines which environment type the feature is in (ie underground, above ground, inside of a building, etc), the load priority determines which features should load first, and the descriptive name is simply a recognizable name that can be used for debugging purposes.

Changed the data path for serializers to the engine tools gui data folder when the application is run in debug mode.

Changed how serializable features were being stored to now use the same internal id setup as everything else. To load the origin, name, id, and load radius without requiring the feature to be loaded into memory a special FeatureList file has been created that stores all of that information for each feature in a single file.

Cleaned up how tile maps were being loaded a bit to make the visuals less jarring.

Added a dev UI for the feature manager. This allows me to see which features are currently loaded and also access a list of all features. Then by selecting a feature from the list of all features I can jump to that location on the map. This UI should make developing new features much easier in the future since I can load and unload features created in the engine tools without restarting the engine client.