Toyed around with the line intersection code. Ended up eventually just implementing a completely new algorithm that works way better than the previous one I'd settled on.
Added support for a forest bounding point id and a river bounding point id.

Added some handling to the ScrollableArea component to make it optionally not scrollable. When used like that it essentially just becomes a scissored area where you can nest more UI elements.

Refactored all event flags (Clickable, Hoverable, etc) into getters and setters so that the UIManager can be dynamically updated when these properties change.

Changed all instances of OnClickAction to use the Click event instead. (OnClickAction was some first week of engine development code so it really had to go)

Added a TileClicked event to Scenes

Added an AnimationManager that loads all of the animation sets into a dictionary so that they can be easily accessed. Loading all of these up front is very cheap since they each can only be a couple dozen bytes big

Added some handling to calculate which maps are affected for a set of bounding points. How this works is it first checks which maps have bounding points in them and then creates a bounding box out of the most extreme map coordinates in each direction. Each map in that bounding box is then sampled every couple of tiles. If a tile is found to be within the bounds the map is added to the affected maps hashset

Changed some cases where I was using ConcurrentBags to HashSets and locks. There really isn't a huge reason for doing this beyond keeping implementations uniform throughout the program.

Changed the entity manager to use the new UnitCreationInfo system instead of the UnitProfile system.

Added a simple world map. This takes each feature that has a map size greater than 0 and displays the specified animation set on the map. When hovering over the feature's UI object the feature name will pop up in a tooltip.