Added support for a new render paradigm for rendering large amount of objects. This involves generating all of the necessary render data and loading it into gpu buffers beforehand then only using the gpu buffers when rendering. Currently this is set up for structures but is planned to replace tile map textures.
Added a secondary render pipeline for fading everything except the UI. This can be used as an example for how to set up alternate render pipelines/post processing in the future.

Changed how ticking worked to now be on a subscription basis only. This improves performance substantially by not requiring every object to be looped through to every time the game needs to tick.

Added some dictionary extension methods. These simplify a couple common operations such as "set this key to this value regardless of whether it already exists" or "get an item from the dictionary or add a default item and return that if the key wasn't found"

Added another item to the FeatureLedgerNode called HashData. This allows string values to be assigned to any given object hash of a feature. Currently this is being used to track whether any given unit in a feature is alive or dead.

Added lua scripting and replaced all cases where state values were being used externally to now be created by calling scripts that set the state values.
Ex: "SubscribeToFeatureTrigger(2, "PlayerInside", "StartQuest(0)")". This will trigger the quest with id 0 when the player enters the bounds of the feature with id 2. This is exponentially simpler to use than having to remember a boatload of arbitrary state values.

Added a new StateSubscriber class that simplifies the subscription handling somewhat. This class also supports applying scripts as the triggered action.

Added a Parameters field and a ClearParameters field to feature equations. Parameters can apply any number of effects based on the values set in the engine tools and ClearParameters are used to determine when the feature has been cleared.

Added support for creating buildings via feature equations.

Added Enter and Exit events to feature equations. These are calculated every time the player's tile position changes and trigger when the player has been determined to cross a feature bounds threshold

Added support for multiple bounding points per feature.

Moved the FeaturePoint definition into its own file because I was getting claustrophobic.

Updated the SaveState to include the new HashData field.

Added UnitParameters to units. These can be any <string, string> key value pair and will be applied to the unit when they are loaded.

Changed the serializable feature's AffectedPoints field to use a new AffectedPoint class that simplifies creation in the tools a bit.

Added a Title and Body field to quests and changed quest objectives to use scripts instead of state values. Ex: "SubscribeToFeatureTrigger(2, "Clear", "CompleteQuestObjective(0, 0, 0)")"

Added a method to easily check if a quest is available (ie it hasn't been completed and it's not in progress). This should be called before any attempted subscriptions to begin a quest.

Removed structures from the scene's unit collation process.

Changed tile map loading to fade out and then back in to hide any jarring load effects.

Updated the documentation to include information about scripting.